HttpStress: Segregate Client & Server into separate classes and support running standalone servers and clients
